pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["MohammadAl-khatib <alk20489@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
black = "^21.12b0"
flake8 = "^4.0.1"
djangorestframework = "^3.12.4"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
